requirements.txt
pymupdf==1.26.4
pillow==11.3.0
